Lista ultima entrega
