+++ BlackjackDemo.java
@@ -1,9 +1,6 @@
 import java.io.Console;
 import java.util.Random;
 import java.util.Scanner;
-
-import javax.lang.model.util.ElementScanner14;
-
 import java.awt.Font;
 
 public class BlackjackDemo {
@@ -27,6 +24,8 @@ public class BlackjackDemo {
         private Carta() {
         }
     }
+
+    // Metodos para obtener informacion de las cartas que se repartan aleatoriamente
 
     /**
 	 * Metodo con el que se obtendra el palo de cualquier carta
@@ -94,23 +93,6 @@ public class BlackjackDemo {
     }
 
     /**
-	 * Metodo con el que entregara se reparten las cartas al azar
-     * para la mano de cartas del crupier o del jugador
-     */
-    //@ requires mazo.length == 56;
-	//@ requires manoCarta.length == 17 || manoCarta.length == 21;
-    //@ ensures \result != null;
-    //@ ensures \result.length == 17 || \result.length == 21;
-	public static Carta[] repartirCartasIniciales(Carta[] manoCarta, Carta[] mazo) {
-		int i = 0;
-		while (i < 2) {
-			manoCarta[i] = darCarta(mazo);
-			i = i + 1;
-		}
-		return manoCarta;
-	}
-
-    /**
 	 * Metodo que permite registrar el monto de la apuesta del jugador
      * para la mano actual
      */
@@ -120,21 +102,15 @@ public class BlackjackDemo {
     public static /*@ non_null */ int obtenerApuesta(int credito, String nombre) {
         int i = 0;
         Scanner entrada = new Scanner(System.in);
-        System.out.println("");
-		// Se le explica el sistema de apuestas y creditos al jugador
-		System.out.println("Siempre que apuestes, se te va descontar la apuesta de tus creditos actuales mientras se este jugando la mano.");
-		System.out.println("Si ganas se te devuelve la apuesta junto con su ganacia.");
-		System.out.println("Si pierdes no se te devolvera la apuesta y terminaras perdiendo dinero.");
-        System.out.println("");
         System.out.print(nombre + ", por favor ingresa el monto que te gustaría apostar en esta mano: ");
         int apuesta = entrada.nextInt();
 
         //@ maintaining 0 <= i <= 4;
         //@ maintaining (\forall int k; 0 <= k && k < i; apuesta > credito || apuesta < 10 || apuesta <= credito || apuesta >= 10 );
         //@ decreases 4 - i;
-        while (i < 3) {
-            if (i == 2) {
-                System.out.println("Se le ha establecido la apuesta minima : 10 creditos");
+        while (i < 4) {
+            if (i == 3) {
+                System.out.println("Se le ha establecido la apuesta minima : 10");
                 apuesta = 10;
                 break;
             } else if (apuesta > credito || apuesta < 10) {
@@ -146,27 +122,32 @@ public class BlackjackDemo {
                 break;
             }
         }
+		System.out.println("Siempre que apuestes, se te va descontar la apuesta de tus creditos actuales mientras se este jugando la mano.");
+		System.out.println("Si ganas se te devuelve la apuesta junto con su ganacia.");
+		System.out.println("Si pierdes no se te devolvera la apuesta y terminaras perdiendo dinero.");
         return apuesta;
     }
+
 
     /**
 	 * Metodo que permite calcular el valor de la suma de las cartas de una mano
      */
-    //@ requires manoCartas.length == 17 || manoCartas.length == 21;
-    //@ requires 2 <= numCartasMano <= manoCartas.length;
-    //@ ensures 2 <= \result || \result <= 31;
-    public static /*@ pure @*/ int valorMano(/*@ non_null */ Carta /*@ non_null */[] manoCartas, int numCartasMano) {
-        int puntosMano = 0;
-        //@ maintaining 0 <= numCartasMano <= manoCartas.length;
-        //@ maintaining (\forall int k; 0 <= k && k < numCartasMano; manoCartas[k].ordinal() % 14 >= 1 || manoCartas[k].ordinal() % 14 <= 11);
+    //@ requires 0 <= numCartasMano <= manoJugador.length;
+    //@ requires 0 <= puntosMano < Integer.MAX_VALUE;
+    //@ ensures 2 <= \result || \result < 31 || \result == puntosMano ;
+    public static /*@ pure @*/ int valorMano(/*@ non_null */ Carta /*@ non_null */[] manoJugador, int numCartasMano, int puntosMano) {
+        puntosMano = 0;
+
+        //@ maintaining 0 <= numCartasMano <= manoJugador.length;
+        //@ maintaining (\forall int k; 0 <= k && k < numCartasMano; manoJugador[k].ordinal() % 14 >= 1 || manoJugador[k].ordinal() % 14 <= 11);
         //@ decreases numCartasMano;
         while (numCartasMano != 0) {
             numCartasMano -= 1;
             //@ assume 0 <= numCartasMano <= 21;
-            //@ assume 0 <= puntosMano <= 31;
-            if (manoCartas[numCartasMano].ordinal() % 14 <= 9)
-                puntosMano = (manoCartas[numCartasMano].ordinal() % 14) + 1 + puntosMano;
-            else if (manoCartas[numCartasMano].ordinal() % 14 <= 12)
+            //@ assume 0 <= puntosMano < 31;
+            if (manoJugador[numCartasMano].ordinal() % 14 <= 9)
+                puntosMano = (manoJugador[numCartasMano].ordinal() % 14) + 1 + puntosMano;
+            else if (manoJugador[numCartasMano].ordinal() % 14 <= 12)
                 puntosMano = 10 + puntosMano ;
             else
                 puntosMano = 11 + puntosMano;
@@ -176,442 +157,478 @@ public class BlackjackDemo {
         return puntosMano;
     }
 
-    /**
-	 * Metodo que dobla la apuesta del jugador
+	/**
+	 * Metodo que permite sacar la ganancia del jugador
      */
-    //@ requires 10 <= credito - apuesta <= 10000;
-    //@ ensures \result == 2 * apuesta;
-	public static /*@ pure @*/ int doblarApuesta(int apuesta, int credito) {
-        apuesta = 2 * apuesta;
-        System.out.println("");
-        System.out.println("Se ha doblado con exito su apuesta.");
-        System.out.println("Ahora su apuesta es de: " + apuesta + " creditos.");
-        System.out.println("Y el numero de creditos totales que posee ahora es: " + credito + ".");
-        return apuesta;
-	}
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 0 <= indicador <= 4;
+    //@ ensures \result == 0 || \result == (apuesta*2) || \result == (apuesta + (int) apuesta * 3/2) || \result == apuesta;
+	public static /*@ pure @*/ int ganancia(int apuesta, int credito, int indicador) {
+		// Ganancia del jugador 
+		int ganancia = 0;
+		if (indicador == 1){
+			// Si hizo blackjack, relacion 3:2
+            //@ assume ganancia == apuesta*3/2;
+			ganancia = apuesta * 3/2;
+            ganancia = (int)ganancia;
+            //@ assume ganancia == ganancia + apuesta;
+            ganancia = ganancia + apuesta;
+			return ganancia;
+		} else if (indicador == 2) {
+			// Gana con una relacion 1:1
+            //@ assume ganancia == apuesta*2;
+			ganancia = apuesta*2;
+			return ganancia;
+		} else if (indicador == 3){
+			// Empate y se le devuelve el dinero
+            //@ assume ganancia == apuesta;
+			ganancia = apuesta;
+			return ganancia;
+		}else 
+			// Si pierde la partida
+            //@ assume ganancia == 0;
+			return ganancia;
+	}  
 
-    /**
-	 * Metodo con el que se le consulta al jugador si desea pedir
-     * una carta adicional, plantarse, doblar su apuesta (cuando
-     * sea posible) o si desea salir del juego
+	/**
+	 * Metodo que muestra las opciones del jugador
      */
 	//@ requires nombre != null;
     //@ ensures \result == 1 || \result == 2 || \result == 3 || \result == 4;
-	public static int consultarDecisionJugador(String nombre, int puntosJugador, boolean acabaDeEmpezarLaMano,
-            int credito, int apuesta) {
+	public static int opciones(String nombre) {
 		// Opciones del jugador
         int k = 0;
 		Scanner entrada = new Scanner(System.in);
-		System.out.println(nombre + ", estas son tus opciones:");
-		System.out.println("(1) Recibir una carta.");
-		System.out.println("(2) Plantarte.");
-		System.out.println("(3) Doblar el monto de la apuesta (si el valor de tus dos primeras cartas es 9, 10 o 11).");
-        System.out.println("Solo se puede doblar el monto de la apuesta si se tienen dos cartas en la mano.");
-		System.out.println("(4) Salir del juego.");
-		System.out.print("¿Que decides?: ");
-		int decision = entrada.nextInt();
+		System.out.println(nombre+ " estas son tus opciones: ");
+		System.out.println("(1) Recibir una carta");
+		System.out.println("(2) Plantarse");
+		System.out.println("(3) Doblar el monto de la apuesta");
+		System.out.println("(4) Salir del juego");
+		System.out.print("Opcion: ");
+		int opcion = entrada.nextInt();
 
-        //@ maintaining 0 <= k <= 2;
-        //@ decreases 3 - k;
-		while (k < 3) {
-            if (k == 2) {
-                System.out.print("Has alcanzado la cantidad maxima de intentos.");
-                System.out.print("Por lo tanto, se seleccionara que se ha plantado.");
-                decision = 2;
-            } else if (decision > 0 && decision < 5 && decision != 3)
+        //@ maintaining 0 <= k <= 4;
+        //@ maintaining (\forall int p; 0 <= p && p < k; opcion != 1 || opcion != 2 ||opcion != 3 ||opcion != 4 || opcion == 1 ||opcion == 2 ||opcion == 3 ||opcion == 4);
+        //@ decreases 4 - k;
+		while (k <= 4) {
+            if (k == 4){
+                System.out.print("Haz alcanzado la cantidad maxima de intentos");
+                System.out.print("Por lo tanto se te seleccionara la opcion 2");
+                opcion = 2;
+                break;
+            } else if (opcion > 0 && opcion < 5) {
 				break;
-            else if (decision == 3) {
-                if (puntosJugador >= 9 && puntosJugador <= 11 && acabaDeEmpezarLaMano == true && (credito - apuesta) >= 10)
-                        break;
-                else {
-                    System.out.println("");
-                    System.out.println("No se puede doblar la apuesta si no se cumplen las condiciones necesarias.");
-                    System.out.println("Para doblar su apuesta, el valor de su mano debe ser 9, 10 u 11.");
-                    System.out.println("Ademas, solo se puede doblar el monto de la apuesta si se tienen dos cartas en la mano.");
-                    System.out.println("Y si se empezo la mano con, al menos, el doble de creditos de la apuesta actual.");
-                    System.out.println("");
-                    System.out.print("Por favor, escoga otra opcion:");
-                    decision = entrada.nextInt();
-                }
-            } else {
-				System.out.print(nombre + " has elegido una opcion erronea.");
-				System.out.print("Por favor, escoga otra opcion:");
-				decision = entrada.nextInt();
+			} else {	
+				System.out.print(nombre +" has elegido una opcion erronea");
+				System.out.print("Por favor escoga una opcion :");
+				opcion = entrada.nextInt();
 			}
-			k = k + 1;
+			k++;
 		}
-		return decision;
-	}
+		return opcion;
+	} 
 
 	/**
 	 * Metodo que pide al jugador el nombre y presenta el juego
      */
 	//@ requires true;
     //@ ensures \result != null;
-	public static String mensajeBienvenida() {
+	public static String mensajeEntrada() {
 		Scanner entrada = new Scanner(System.in);
-		// Se le da al jugador un energico saludo de bienvenida
 		System.out.println("¡Desafia a la computadora en Blackjack!");
         System.out.println("Desarrollado por Isea, Luis (19-10175) y Prieto, Jesus (19-10211).");
         System.out.println("");
-
-		// Se le pide el nombre al usuario antes de cualquier cosa
         System.out.print("Por favor, ingrese su nombre: ");
 		String nombre = entrada.nextLine();
-
-		// Se saluda al jugador con su nombre
         System.out.println("Es un placer tenerte aqui, " + nombre + ".");
         System.out.println("");
-
-		// Se le habla un poco del sistema de creditos
-		System.out.println("Posees 100 creditos por empezar a jugar. La apuesta minima por mano son 10 creditos.");
 		return nombre;
 	}
-
+	
 	/**
-	 * Metodo con el que se muestra el valor de sus cartas y la del crupier al jugador
-     * a traves de la interfaz de texto
+	 * Metodo que le muestra el valor de sus cartas y la del crupier al jugador
      */
-	//@ requires 1 <= puntosCrupier <= 31;
-	//@ requires 2 <= puntosJugador <= 31;
-	public static /*@ pure @*/ void mostrarPuntuacionesPorTexto(int puntosCrupier, int puntosJugador, boolean mostrarCartasCrupier,
-            Carta[] crupier) {
+	//@ requires 1 <= puntosJugador <= 21;
+    //@ ensures true;
+	public static void primeraMuestra(int puntosJugador, Carta crupier[]) {
 		System.out.println("");
-		if (mostrarCartasCrupier == false) {
-            puntosCrupier = obtenerValorCarta(crupier[0]);
-			System.out.println("Valor de la carta descubierta del crupier: " + puntosCrupier);
-        } else
-			System.out.println("Valor de la mano del crupier: " + puntosCrupier);
-        System.out.println("Valor de la mano del jugador: " + puntosJugador);
-		System.out.println("");
-	}
-
-
-	/**
-	 * Metodo con el que se muestra la cantidad de creditos que posee el jugador
-     * y la cantidad de creditos apostados en la mano actual, a traves
-     * de la interfaz de texto
-     */
-	//@ requires 0 <= credito <= 1000000;
-	//@ requires 10 <= apuesta <= 10000;
-	public static /*@ pure @*/ void mostrarCreditosPorTexto(int credito, int apuesta) {
-        System.out.println("Total de creditos restantes: " + credito);
-        System.out.println("Creditos apostados: " + apuesta);
+        //@ assume (\exists int i; 0 <= i && i < 56; crupier[0] == Carta.values()[i]);
+        //@ assume (\exists int j; 0 < j && j <= 11;obtenerValorCarta(crupier[0]) == j);
+        //@ assume crupier.length == 17 ;
+        //@ assume 0 < obtenerValorCarta(crupier[0]) <= 11;
+		System.out.println("Valor de la carta descubierta del crupier: " + obtenerValorCarta(crupier[0]));
+        System.out.println("Valor de la mano actual: " + puntosJugador);
 		System.out.println("");
 	}
 
 	/**
-	 * Metodo con el que se calcula como queda el credito disponible
-     * del jugador luego de apostar, doblar su apuesta o de ganar
-     * una mano
+	 * Metodo que le muestra el mensaje al jugador si perdio, empato o hizo blackjack
      */
-	//@ requires 10 <= apuesta <= 100000;
-    //@ requires 0 <= credito <= 100000;
-	//@ requires 0 <= operacion <= 3;
-    //@ ensures \result == (credito - apuesta) || \result == (credito + (2 * apuesta)) || \result == credito + apuesta + ((apuesta * 3) / 2) || \result == credito + apuesta;
-	public static /*@ pure @*/ int calcularCreditoResultante(int apuesta, int credito, int operacion) {
-		// El jugador acaba de elegir el monto de la apuesta para la mano
-        // o tambien puede que el jugador acaba de doblar su apuesta
-        if (operacion == 0)
-            //@ assume 0 <= credito - apuesta <= 100000;
+	//@ requires 1 <= puntosJugador <= 33;
+	//@ requires 1 <= puntosCrupier <= 21;
+    //@ ensures \result == 1 || \result == 2 || \result == 3 || \result == 4;
+	public static int mensajeJugador(int puntosCrupier, int puntosJugador) {
+		// Mensaje para el jugador en cada caso
+		if (puntosJugador > 21){
+			System.out.println("Haz perdido la partida.");
+			System.out.println("Tu puntaje es mayor que 21.");
+			return 1;
+		} else if (puntosJugador == puntosCrupier) {
+			System.out.println("Ha ocurrido un empate");
+			System.out.println("Tienes los mismo puntos que el crupier");
+			return 2;
+		} else if (puntosJugador == 21) {
+			System.out.println("¡Hiciste BLACKJACK!");
+			System.out.println("Tu puntaje es de 21.");
+			System.out.println("Haz ganado la partida.");
+			return 3;
+		}else
+			return 4;
+	}
+
+	/**
+	 * Metodo que muestra si el crupier perdio 
+     */
+	//@ requires 1 <= puntosCrupier <= 33;
+    //@ ensures \result == 1 || \result == 2 ;
+	public static int mensajeCrupier(int puntosCrupier ) {
+		// Mensaje del crupier si se paso del limite
+		if (puntosCrupier > 21){
+			System.out.println("El puntaje del crupier es mayor que 21");
+			System.out.println("Haz ganado la partida.");
+			return 1;
+		}else
+			return 2;
+	}
+
+	 /**
+	 * Metodo que verifica si puede doblar la apuesta y la dobla
+     */
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 1 <= i <= 21;
+	//@ requires 1 <= puntosJugador <= 33;
+	//@ requires 1 <= puntosCrupier <= 33;
+    //@ ensures \result == apuesta*2 || \result == apuesta;
+	public static int doblarApuesta(int puntosJugador, int puntosCrupier, int apuesta, int credito, int i) {
+		// Verificar si se puede doblar la apuesta
+        //@ assume apuesta*2 < credito;
+		if ((puntosJugador == 9 || puntosJugador == 10 || puntosJugador == 11) && apuesta*2 < credito && i == 2) {
+			// Doblarla
+            //@ assume apuesta == apuesta*2;
+			apuesta = apuesta*2;
+			return apuesta;
+		}else { 
+			// Mensaje de error al no cumplir verificacion
+            //@ assume apuesta == apuesta;
+			System.out.println("Para doblar necesita que la suma de sus dos primeras cartas sea 9,10 o 11 y tener suficiente credito");
+			System.out.println("Como no cumple el requisito no puede doblar la apuesta.");
+			return apuesta;
+		}
+	}
+
+	/**
+	 * Metodo que registra la resta del credito
+     */
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 0 <= indicador <= 4;
+    //@ ensures \result == (credito-apuesta) || \result == (credito -(apuesta/2)) || \result == credito;
+	public static int operacionCredito(int apuesta, int credito, int indicador) {
+		if (indicador == 1){
+			// Restar credito
+            //@ assume credito == credito-apuesta;
 			credito = credito - apuesta;
-        // El jugador hizo blackjack y recibe su ganancia con razon
-        // de 3 a 2 con respecto a su apuesta
-		else if (operacion == 1)
-            //@ assume 0 <= credito + apuesta + ((apuesta * 3) / 2) <= 100000;
-            credito = credito + apuesta + ((apuesta * 3) / 2);
-        // El jugador gano la mano con una ganancia con razon de 1 a 1
-        else if (operacion == 2)
-            //@ assume 0 <= credito + (2 * apuesta) <= 100000;
-            credito = credito + (2 * apuesta);
-        // El jugador empato la mano con el crupier, se le
-        // devuelve la apuesta
-        else
-            //@ assume 0 <= credito + apuesta <= 100000;
-            credito = credito + apuesta;
+			return credito;	
+		}else if (indicador == 3){
+			// Restar credito al decidir apostar el doble
+            //@ assume apuesta == apuesta/2;
+			apuesta = apuesta/2;
+            //@ assume credito == credito-apuesta;
+            credito = credito - apuesta;
+			return credito;
+		} else
+            //@ assume credito == credito;
+			return credito;
+	}
+
+	/**
+	 * Metodo que verifica si el jugador hizo blackjack
+     */
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 0 <= mensaje <= 5;
+	//@ requires 1 <= puntosJugador <= 33;
+	//@ requires 1 <= puntosCrupier <= 33;
+    //@ ensures credito <= \result <= credito + (apuesta + apuesta *(3/2));
+	public static int ver_blackjack(int apuesta, int credito, int puntosCrupier, int puntosJugador, int mensaje) {
+		//Verificar si tiene BLACKJACK o empato
+		if (mensaje == 2){
+			// Hubo empate
+            //@ assume credito == credito + ganancia(apuesta, credito, 3);
+			credito = credito + ganancia(apuesta, credito, 3);
+			return credito;
+		} else if (mensaje == 3){
+            //Hizo BLACKJACK
+            //@ assume credito == credito + ganancia(apuesta, credito, 1);
+			credito = credito + ganancia(apuesta, credito, 1);
+			return credito;
+		}else 
+			// No ocurrio nada
+            //@ assume credito == credito;
+			return credito;
+	}
+
+	/**
+	 * Metodo que verifica si el jugador hizo blackjack, empato o perdio
+     */
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 0 <= mensaje <= 5;
+	//@ requires 1 <= puntosJugador <= 33;
+	//@ requires 1 <= puntosCrupier <= 33;
+    //@ ensures credito <= \result <= credito + (apuesta + apuesta *(3/2));
+	public static int verificacion(int apuesta, int credito, int puntosCrupier, int puntosJugador, int mensaje) {
+		//Verificar si tiene BLACKJACK o empato
+		if (puntosJugador > 21)
+			mensaje = 1;
+		else if (puntosJugador == puntosCrupier) 
+			mensaje = 2;
+		else if (puntosJugador == 21) 
+			mensaje = 3;
+		else
+			mensaje = 4;
+
+		if (mensaje == 2) {
+            // EMPATE
+            //@ assume credito == credito + ganancia(apuesta, credito, 3);
+			credito = credito + ganancia(apuesta, credito, 3);
+			return credito;
+		} else if (mensaje == 3) {
+            // BLACKJACK
+            //@ assume credito == credito + ganancia(apuesta, credito, 1);
+			credito = credito + ganancia(apuesta, credito, 1);
+			return credito;
+		} else
+			//@ assume credito == credito;
+			return credito;
+	}
+	
+	
+	/**
+	 * Metodo que verifica y genera la ganacia del jugador si el crupier perdio
+     */
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 0 <= mensaje <= 5;
+	//@ requires 1 <= puntosJugador <= 33;
+	//@ requires 1 <= puntosCrupier <= 33;
+    //@ ensures credito <= \result <= credito + (apuesta + apuesta *(3/2));
+	public static int ver_crupier(int apuesta, int credito, int puntosCrupier, int puntosJugador, int mensaje) {
+		if (mensaje == 1){
+			//Si el crupier se paso de 21
+            //@ assume credito == credito + ganancia(apuesta, credito, 2);
+			credito = credito + ganancia(apuesta, credito, 2);
+			return credito;
+		} else 
+			// Si el crupier se mantuvo
+            //@ assume credito == credito;
+			return credito;
+	}
+
+	/**
+	 * Metodo que verifica los puntos del Crupier y el jugador y toma una decision
+     */
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 0 <= mensaje <= 5;
+	//@ requires 1 <= puntosJugador <= 33;
+	//@ requires 1 <= puntosCrupier <= 33;
+    //@ ensures credito <= \result <= credito + (apuesta + apuesta *(3/2));
+	public static int decision(int apuesta, int credito, int puntosCrupier, int puntosJugador, int mensaje) {
+		// Mensaje para el jugador en cada caso
+		if (puntosJugador < puntosCrupier){
+			System.out.println("Haz perdido la partida.");
+			mensaje = 1;
+		}else if (puntosJugador > puntosCrupier) {
+			System.out.println("Haz ganado la partida.");
+			mensaje = 2;
+		} else if (puntosJugador == puntosCrupier) {
+			System.out.println("Ha ocurrido un empate");
+			mensaje = 3;
+		} else
+			mensaje = 4;
+
+		// Obtener credito con respecto a la decision
+		if (mensaje == 3) {
+			// Retornar la apuesta, hubo empate
+            //@ assume credito == credito + ganancia(apuesta, credito, 3);
+			credito = credito + ganancia(apuesta, credito, 3);
+			return credito; 
+		} else if (mensaje == 2) {
+			// Jugador gana con una relacion 1:1
+            //@ assume credito == credito + ganancia(apuesta, credito, 2);
+			credito = credito + ganancia(apuesta, credito, 2);
+			return credito;
+		} else if (mensaje == 1) {
+			// Jugador pierde y no se le devuelve le dinero
+            //@ assume credito == credito + ganancia(apuesta, credito, 4);
+			credito = credito + ganancia(apuesta, credito, 4);
+			return credito;
+			//actualizar estado y componente grafico
+		} else 
+            //@ assume credito == credito;
+			return credito;
+	}
+	
+	/**
+	 * Metodo que verifica los puntos del jugador 
+     */
+	//@ requires Integer.MIN_VALUE <= cota_FM <= Integer.MAX_VALUE;
+    //@ requires 0 <= credito < Integer.MAX_VALUE;
+    //@ ensures \result == 1 || \result == 0;
+	public static int end_mano(int cota_FM, int credito) {
+        int k = 0;
+		Scanner entrada = new Scanner(System.in);
+		System.out.println("Tu credito actual: "+ credito);
+		System.out.println("¿Deseas jugar otro mano?");
+		System.out.println("Para SI escribe 1");
+		System.out.println("Para NO escribe 0");
+		System.out.print("Y/N:");
+		cota_FM = entrada.nextInt();
+        //@ maintaining 0 <= k <= 4;
+        //@ maintaining (\forall int p; 0 <= p && p < k; cota_FM != 1 || cota_FM != 0 || cota_FM == 1 || cota_FM == 0 );
+        //@ decreases 4 - k;
+		while (k <= 4) {
+            if (k == 4) {
+                System.out.println("Haz alcanzado la cantidad maxima de intentos");
+                System.out.println("Por lo tanto tomares que ya no deseas jugar");
+                cota_FM = 0;
+                break;
+            } else if (cota_FM == 0 || cota_FM == 1) {
+                break;
+            } else {
+                System.out.println("Ha ingresado un caracter erroneo");
+                System.out.print("Por favor ingrese 1 o 0:");
+                cota_FM = entrada.nextInt();
+            }
+            k++;
+		}
+		return cota_FM;
+	}
+
+	/**
+	 * Metodo que entregara las dos primeras cartas
+     */
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 0 <= mensaje <= 5;
+	//@ requires 1 <= puntosJugador <= 33;
+	//@ requires 1 <= puntosCrupier <= 33;
+    //@ ensures credito <= \result <= credito + (apuesta + apuesta *(3/2));
+	public static Carta[] primeras2Cartas(Carta[] baraja, Carta[] mazo) {
+		int i = 0;
+		
+		while (i <= 2) {
+			if(i == 2)
+				break;
+			baraja[i] = darCarta(mazo);
+			i++;
+		}
+		return baraja;
+	}
+
+	/**
+	 * Metodo que entregara las dos primeras cartas
+     */
+	//@ requires 10 <= apuesta <= credito;
+    //@ requires 10 <= credito < Integer.MAX_VALUE;
+	//@ requires 0 <= mensaje <= 5;
+	//@ requires 1 <= puntosJugador <= 33;
+	//@ requires 1 <= puntosCrupier <= 33;
+    //@ ensures credito <= \result <= credito + (apuesta + apuesta *(3/2));
+	public static int acciones(int puntosJugador, int puntosCrupier, int mensaje, int mensajeJugador, int apuesta, int credito, int opcion, int comprobante, Carta[] mazo, Carta[] jugador, Carta[] crupier, String nombre) {
+		int i = 2;
+		int j = 2;
+		//Accion del jugador 
+		mensaje = mensajeJugador(puntosJugador, puntosCrupier);
+		credito = ver_blackjack(apuesta, credito, puntosCrupier, puntosJugador, mensaje);
+		if (mensaje == 2 || mensaje == 3)
+			return credito; 	
+			
+		while (puntosJugador <= 21) {
+			opcion = opciones(nombre);
+			while (opcion == 1){
+				// el jugador pide una carta
+				jugador[i] = darCarta(mazo);
+				i++;
+				// Calculo del puntaje
+				puntosJugador = valorMano(jugador, i, puntosJugador);
+				System.out.println("Este es tu puntaje: "+ puntosJugador);
+				// Verificion del puntaje
+				credito = verificacion(apuesta, credito, puntosCrupier, puntosJugador, mensaje);
+				if (mensajeJugador(puntosCrupier, puntosJugador) != 4)
+					return credito;
+				opcion = opciones(nombre);
+			}
+			if (opcion == 2) {
+				// El jugador se planta
+				break;
+				// Actualizar estado
+			}else if (opcion == 3) {
+				// El jugador dobla la apuesta
+				comprobante = doblarApuesta(puntosJugador, puntosCrupier, apuesta, credito, i);
+				if (comprobante != apuesta ) {
+					apuesta = comprobante;
+					credito = operacionCredito(apuesta, credito, 3);
+					jugador[i] = darCarta(mazo);
+					i++;
+					puntosJugador = valorMano(jugador, i, puntosJugador);
+					System.out.println("Este es tu puntaje: "+ puntosJugador);
+					credito = verificacion(apuesta, credito, puntosCrupier, puntosJugador, mensaje);
+					if (mensajeJugador(puntosJugador, puntosCrupier) != 4) {
+						return credito;
+					}
+				}	
+			}else {
+				// El jugador decide salirse del juego
+				return 0;
+
+			}
+		}
+		
+		if (puntosJugador > 21){
+			//el jugador sobrepaso los 21 
+			mensaje = mensajeJugador(puntosJugador, puntosCrupier);
+			return credito;
+		}
+				
+		//Acciones del crupier
+		while (puntosCrupier < 17) {
+			//dar carta
+			crupier[j] = darCarta(mazo);
+			j++;
+			puntosCrupier = valorMano(crupier, j, puntosCrupier);
+		}
+		if (17 <= puntosCrupier && puntosCrupier <= 21) {
+			// Plantarse
+			// Skip
+			// Actualizar 
+		} else {
+			// presentar perdida del crupier
+			mensaje = mensajeCrupier(puntosCrupier);
+			credito = ver_crupier(apuesta, credito, puntosCrupier, puntosJugador, mensaje);
+			return credito;
+		}
+		//Decision del Juego
+		credito = decision(apuesta, credito, puntosCrupier, puntosJugador, mensaje);
 		return credito;
 	}
 
-    /**
-	 * Metodo que verifica si el jugador o el crupier se pasaron de 21
-     */
-	//@ requires 2 <= puntosJugador <= 31;
-	//@ requires 2 <= puntosCrupier <= 28;
-    //@ ensures \result == (puntosCrupier > 21 || puntosJugador > 21);
-	public static /*@ pure @*/ boolean revisarSiSePasaronDeVeintiuno(int puntosCrupier, int puntosJugador) {
-		boolean sePasaronDeVeintiuno = false;
-        // O el jugador o el crupier se pasaron de 21
-        // Tal vez ambos
-        if (puntosCrupier > 21 || puntosJugador > 21)
-			sePasaronDeVeintiuno = true;
-        return sePasaronDeVeintiuno;
-	}
-
 	/**
-	 * Metodo que verifica si el jugador o el crupier hicieron blackjack
-     */
-	//@ requires 2 <= puntosJugador <= 31;
-	//@ requires 2 <= puntosCrupier <= 28;
-    //@ ensures \result == (puntosCrupier == 21 || puntosJugador == 21);
-	public static /*@ pure @*/ boolean revisarSiHayBlackjack(int puntosCrupier, int puntosJugador) {
-		boolean siHayBlackjack = false;
-        // O el jugador o el crupier hicieron blackjack
-        // Tal vez ambos
-        if (puntosCrupier == 21 || puntosJugador == 21)
-			siHayBlackjack = true;
-        return siHayBlackjack;
-	}
-
-	/**
-	 * Metodo con el que se determina si el jugador gano, empatado o perdio
-	 * la mano de blackjack. Se le proporciona esta informacion al jugador
-     * a traves de la interfaz de texto. El metodo retorna el credito
-     * resultante luego de conocer el resultado de la mano
-     */
-	//@ requires 10 <= apuesta <= 100000;
-    //@ requires 0 <= credito < 100000;
-	//@ requires 2 <= puntosJugador <= 31;
-	//@ requires 2 <= puntosCrupier <= 28;
-    //@ ensures credito <= \result <= credito + ((apuesta + apuesta) * (3 / 2));
-	public static /*@ pure @*/ int determinarResultadoMano(int apuesta, int credito, int puntosCrupier, int puntosJugador,
-            String nombre) {
-        System.out.println("");
-        // Tanto el crupier como el jugador hicieron blackjack
-        // Esto es un empate
-        if (puntosCrupier == 21 && puntosJugador == 21) {
-            System.out.println("AMBOS hicieron BLACKJACK, empataron la mano.");
-            System.out.println("Como ambos obtuvieron 21 puntos, se te devolvera la apuesta a tus creditos disponibles.");
-			credito = calcularCreditoResultante(apuesta, credito, 3);
-        // El crupier hizo blackjack
-        } else if (puntosCrupier == 21) {
-            System.out.println("El CRUPIER hizo BLACKJACK, obtuvo 21 puntos. Perdiste la mano");
-            System.out.println("Lo siento, pero no se te devolvera el dinero de la apuesta.");
-        // El jugador hizo blackjack
-        } else if (puntosJugador == 21) {
-            System.out.println("HICISTE BLACKJACK, obtuviste 21 puntos ¡Ganaste, " + nombre + "!");
-            System.out.println("¡Felicidades! Se te devolvera la apuesta y se te entregara una ganancia en razon 3:2.");
-            credito = calcularCreditoResultante(apuesta, credito, 1);
-
-        // Ninguno de los dos hizo blackjack
-        // Verificar entonces si alguno se paso de 21
-
-        // Tanto el crupier como el jugador se pasaron de 21
-        // Esto es un empate
-        } else if (puntosCrupier > 21 && puntosJugador > 21) {
-            System.out.println("AMBOS se PASARON de 21 puntos, empataron la mano.");
-            System.out.println("Se te devolvera la apuesta a tus creditos disponibles.");
-            credito = calcularCreditoResultante(apuesta, credito, 3);
-        // El crupier se paso de 21
-        } else if (puntosCrupier > 21) {
-            System.out.println("El CRUPIER se PASO de 21 puntos ¡Ganaste, " + nombre + "!");
-            System.out.println("Se te devolvera la apuesta y se te entregara una ganancia en razon 1:1.");
-            credito = calcularCreditoResultante(apuesta, credito, 2);
-        // El jugador se paso de 21
-        } else if (puntosJugador > 21) {
-            System.out.println("Te PASASTE de 21 puntos, perdiste la mano.");
-            System.out.println("Lo siento, pero no se te devolvera el dinero de la apuesta.");
-
-        // Ninguno de los dos se paso de 21
-        // Verificar entonces quien quedo mas cerca de 21
-
-        // Tanto el crupier como el jugador obtuvieron la misma cantidad de puntos
-        // Esto es un empate
-        } else if (puntosCrupier == puntosJugador) {
-            System.out.println("AMBOS obtuvieron la MISMA puntuacion, empataron la mano.");
-            System.out.println("Se te devolvera la apuesta a tus creditos disponibles.");
-            credito = calcularCreditoResultante(apuesta, credito, 3);
-        // El crupier quedo mas cerca de 21 que el jugador
-        } else if (puntosCrupier > puntosJugador) {
-            System.out.println("El CRUPIER quedo MAS CERCA de 21, perdiste la mano.");
-            System.out.println("Lo siento, pero no se te devolvera el dinero de la apuesta.");
-        // El jugador quedo mas de 21 puntos que el crupier
-        } else if (puntosJugador > puntosCrupier) {
-            System.out.println("Quedaste MAS CERCA de 21 ¡Ganaste, " + nombre + "!");
-            System.out.println("Se te devolvera la apuesta y se te entregara una ganancia en razon 1:1.");
-            credito = calcularCreditoResultante(apuesta, credito, 2);
-        }
-
-        return credito;
-	}
-
-	/**
-	 * Metodo con el que se determina si el usuario quiere jugar otra mano.
-	 * Tambien se le avisa al jugador cuando haya llegado al numero
-	 * maximo de manos seguidas que puede jugar por sesion y se termina
-	 * el juego
-     */
-    //@ requires 0 <= credito <= 1000000;
-    //@ ensures \result == true || \result == false;
-	public static /*@ pure @*/ boolean determinarSiSeJugaraOtraMano(int credito) {
-        int k = 0;
-		Scanner entrada = new Scanner(System.in);
-        boolean seJugaraOtraMano = false;
-        System.out.println("");
-		System.out.println("Tu credito actual es: " + credito);
-		System.out.println("¿Deseas jugar otro mano?");
-		System.out.println("Para SI escribe 1. Para NO escribe 0.");
-		System.out.print("1/0: ");
-		int decisionJugador = entrada.nextInt();
-        //@ maintaining 0 <= k <= 3;
-        //@ maintaining (\forall int p; 0 <= p && p < k; decisionJugador != 1 || decisionJugador != 0 || decisionJugador == 1 || decisionJugador == 0);
-        //@ decreases 3 - k;
-		while (k < 3) {
-            if (k == 2) {
-                System.out.print("Has alcanzado la cantidad maxima de intentos.");
-                System.out.print("Por lo tanto, se asumira que ya no desea jugar.");
-                seJugaraOtraMano = false;
-            } else if (decisionJugador == 0) {
-                seJugaraOtraMano = false;
-                break;
-            } else if (decisionJugador == 1) {
-                seJugaraOtraMano = true;
-                break;
-            } else {
-                System.out.println("Ha ingresado un caracter invalido.");
-                System.out.print("Por favor, ingrese 1 para jugar una nueva mano o 0 para no jugar mas: ");
-                decisionJugador = entrada.nextInt();
-            }
-            k = k + 1;
-		}
-		return seJugaraOtraMano;
-	}
-
-	/**
-	 * Metodo con el que se determina si el color del palo y el texto de la carta
-	 * sera negro o rojo. A Picas y a Treboles se les asigna el color negro.
-	 * Mientras que a Diamantas y a Corazones se les asigna el color rojo
-	 */
-	//@ requires (paloCarta.equals("Picas") || paloCarta.equals("Diamantes") || paloCarta.equals("Treboles") || paloCarta.equals("Corazones"));
-	//@ ensures (\result == Colores.BLACK) <== (paloCarta.equals("Picas") || paloCarta.equals("Treboles"));
-	//@ ensures (\result == Colores.RED) <== (paloCarta.equals("Diamantes") || paloCarta.equals("Corazones"));
-	public static /*@ pure @*/ Colores determinarColorCarta(String paloCarta) {
-		// Si el palo es picas o treboles, se asigna el color negro
-		if (paloCarta.equals("Picas") || paloCarta.equals("Treboles"))
-			return Colores.BLACK;
-		// Si el palo es diamantes o corazones, se asigna el color rojo
-		else
-			return Colores.RED;
-	}
-
-	/**
-	 * Metodo con el que se determina cuantas cartas deben ir por fila a la
-	 * hora de dibujarlas en la mesa de BlackJack. El objetivo es aprovechar
-	 * el espacio de la mesa y no agrupar todas las cartas en una sola fila
-	 */
-	//@ requires 0 <= numCartasMano <= 21;
-	//@ requires poseedorMano.equals("crupier") || poseedorMano.equals("jugador");
-	//@ ensures \result != null;
-	//@ ensures \result instanceof int[];
-	//@ ensures \result.length == 3;
-	public static /*@ pure @*/ int[] getSizes(int numCartasMano, String poseedorMano) {
-		int[] sizes = new int[3];
-		if (poseedorMano.equals("crupier")) {
-			if (numCartasMano <= 12)
-				sizes[0] = numCartasMano;
-			else {
-				sizes[0] = 12;
-				sizes[1] = numCartasMano - 12;
-			}
-		} else {
-			if (numCartasMano <= 10)
-				sizes[0] = numCartasMano;
-			else if (numCartasMano <= 18) {
-				sizes[0] = 10;
-				sizes[1] = numCartasMano - 10;
-			} else {
-				sizes[0] = 10;
-				sizes[1] = 8;
-				sizes[2] = numCartasMano - 18;
-			}
-		}
-		return sizes;
-	}
-
-	/**
-	 * Metodo con el que dado el numero de cartas de una mano, se construye un
-	 * arreglo que determina la posicion X de cada una de las cartas
-	 */
-	//@ requires 0 <= numCartasMano <= 21;
-	//@ requires 30 <= anchoCarta <= 10000;
-	//@ requires 0 < anchoMesa < Integer.MAX_VALUE;
-	//@ ensures \result != null;
-	//@ ensures \result instanceof int[];
-	//@ ensures \result.length == numCartasMano;
-	public static /*@ pure @*/ int[] posicionesManoCarta(int numCartasMano, int anchoCarta, int anchoMesa) {
-		int espacio = 20;
-		int bloque = anchoCarta + espacio;
-		int[] posicionesXdeCartas = new int[numCartasMano];
-		int i = 0;
-		//@ maintaining 0 <= i <= numCartasMano;
-		//@ decreases numCartasMano - i;
-		while (i < numCartasMano) {
-			if (numCartasMano % 2 == 0) {
-				//@ assume Integer.MIN_VALUE < i - (numCartasMano / 2) < Integer.MAX_VALUE;
-				int posicion = i - (numCartasMano / 2);
-				//@ assume Integer.MIN_VALUE < (anchoMesa / 2) + (bloque * posicion) + (espacio / 2) < Integer.MAX_VALUE;
-				posicionesXdeCartas[i] = (anchoMesa / 2) + (bloque * posicion) + (espacio / 2);
-			} else {
-				//@ assume Integer.MIN_VALUE < i - ((numCartasMano - 1 )/ 2) < Integer.MAX_VALUE;
-				int posicion = i - ((numCartasMano - 1) / 2);
-				//@ assume Integer.MIN_VALUE < (anchoMesa / 2) + (bloque * posicion) - (anchoCarta / 2) < Integer.MAX_VALUE;
-				posicionesXdeCartas[i] = (anchoMesa / 2) + (bloque * posicion) - (anchoCarta / 2);
-			}
-			i = i + 1;
-		}
-		return posicionesXdeCartas;
-	}
-
-	/**
-	 * Metodo con el que se dibujara una mesa de BlackJack, con un semicirculo verde
-	 * que imite las mesas de BlackJack reales. Tambien se dibujaran las zonas
-	 * que delimitan donde van las cartas del jugador y donde las del crupier
-	 */
-	//@ requires mesa != null;
-	//@ requires 0 < anchoMesa <= mesa.XMAX < Integer.MAX_VALUE;
-	//@ requires 0 < alturaMesa <= mesa.YMAX < Integer.MAX_VALUE;
-	//@ requires 30 <= alturaCarta <= 10000;
-	//@ requires 30 <= anchoCarta <= 10000;
-	public static /*@ pure @*/ void dibujarMesaBlackjack(MaquinaDeTrazados mesa, int anchoMesa, int alturaMesa, int alturaCarta,
-			int anchoCarta) {
-		// Semicírculo verde que imita una mesa de BlackJack real con un borde oscuro
-		mesa.dibujarOvaloLleno(1, -(anchoMesa / 2), anchoMesa, anchoMesa, Colores.DARK_GRAY);
-		mesa.dibujarOvaloLleno(9, -(anchoMesa / 2), anchoMesa - 16, anchoMesa - 16, Colores.GREEN);
-
-		// Dibujar la zona que delimita donde van las cartas del jugador
-		mesa.dibujarOvalo(57, -(anchoMesa / 2), anchoMesa - 112, anchoMesa - 20, Colores.RED);
-		int[] xPuntos = new int[] {35, 70 , anchoMesa - 35, anchoMesa - 70};
-		int[] yPuntos = new int[] {-5, 258 , -5, 258};
-		mesa.dibujarPoligonoLleno(xPuntos, yPuntos, 4, Colores.GREEN);
-		mesa.dibujarLinea(110, 250, anchoMesa - 108, 250, Colores.RED);
-		mesa.configurarFuente("SansSerif", Font.BOLD, 24);
-		mesa.dibujarString("Jugador", 385, 535, Colores.RED);
-		mesa.dibujarString("Jugador", anchoMesa - 487, 535, Colores.RED);
-
-		// Dibujar la zona que delimita donde van las cartas del crupier
-		int[] xPuntos2 = new int[] {85, 85 , 405, 405, anchoMesa - 405, anchoMesa - 405, anchoMesa - 85, anchoMesa - 85};
-		int[] yPuntos2 = new int[] {5, 123, 123, 244, 244, 123, 123, 5};
-		mesa.dibujarPoligono(xPuntos2, yPuntos2, 8, Colores.BLACK);
-		mesa.dibujarString("Crupier", 280, 192, Colores.BLACK);
-		mesa.dibujarString("Crupier", anchoMesa - 385, 192, Colores.BLACK);
-
-		// Escribir BlackJack debajo del semicirculo verde
-		mesa.configurarFuente("SansSerif", Font.BOLD, 38);
-		mesa.dibujarString("Black", (anchoMesa / 2) - 110, alturaMesa - 95, Colores.BLACK);
-		mesa.dibujarString("Jack", (anchoMesa / 2) + 10, alturaMesa - 95, Colores.RED);
-
-		// Dibujar los cuatro palos de las cartas de blackjack centrados debajo del BlackJack central
-		//@ assume 0 < (anchoMesa / 2) - 110 < Integer.MAX_VALUE;
-		//@ assume 0 < alturaMesa - 100 < Integer.MAX_VALUE;
-		//@ assume alturaMesa - 100 + alturaCarta < Integer.MAX_VALUE;
-		dibujarPaloCarta(mesa, (anchoMesa / 2) - 110, alturaMesa - 115, alturaCarta, anchoCarta, "Picas", Colores.BLACK);
-		dibujarPaloCarta(mesa, (anchoMesa / 2) - 65, alturaMesa - 115, alturaCarta, anchoCarta, "Diamantes", Colores.RED);
-		dibujarPaloCarta(mesa, (anchoMesa / 2) - 20, alturaMesa - 115, alturaCarta, anchoCarta, "Treboles", Colores.BLACK);
-		dibujarPaloCarta(mesa, (anchoMesa / 2) + 25, alturaMesa - 115, alturaCarta, anchoCarta, "Corazones", Colores.RED);
-	}
-
-	/**
-	 * Metodo con el que se dibuja el simbolo del palo de la carta.
-	 * El simbolo que se dibujara estara centrado
+	 * Metodo con el que se mostrara graficamente las cartas visibles que vayan
+	 * apareciendo en la mano del jugador y en la del crupier
 	 */
 	//@ requires mesa != null;
 	//@ requires 0 < mesa.XMAX < Integer.MAX_VALUE;
@@ -625,9 +642,14 @@ public class BlackjackDemo {
 	//@ requires (posXdeCarta + (anchoCarta / 2) + 25) < Integer.MAX_VALUE;
 	//@ requires (posYdeCarta + (alturaCarta / 2) + 18) < Integer.MAX_VALUE;
 	//@ requires paloCarta.equals("Picas") || paloCarta.equals("Diamantes") || paloCarta.equals("Treboles") || paloCarta.equals("Corazones");
-	//@ requires color == Colores.BLACK || color == Colores.RED;
-	public static /*@ pure @*/ void dibujarPaloCarta(MaquinaDeTrazados mesa, int posXdeCarta, int posYdeCarta,
-			int alturaCarta, int anchoCarta, String paloCarta, Colores color) {
+	//@ requires nombreCarta.length() > 0;
+	public static void dibujarCartaVisible(MaquinaDeTrazados mesa, int posXdeCarta, int posYdeCarta,
+			int alturaCarta, int anchoCarta, String paloCarta, String nombreCarta) {
+		// Dibujar el rectángulo externo e interno de la carta
+		mesa.dibujarRectanguloLleno(posXdeCarta, posYdeCarta, 72, 108, Colores.WHITE);
+		mesa.dibujarRectangulo(posXdeCarta + 4, posYdeCarta + 5, anchoCarta - 10, alturaCarta - 10, Colores.DARK_GRAY);
+		Colores color = Colores.BLACK;
+
 		// Dibujar el símbolo de los palos de las cartas
 		if (paloCarta.equals("Picas")) {
 			// Dibujar el símbolo de Picas
@@ -648,6 +670,7 @@ public class BlackjackDemo {
 
 		} else if (paloCarta.equals("Diamantes")) {
 			// Dibujar el símbolo de Diamantes
+			color = Colores.RED;
 			int[] xPuntos = new int[] { posXdeCarta + (anchoCarta / 2) - 12, posXdeCarta + (anchoCarta / 2),
 					posXdeCarta + (anchoCarta / 2) + 12, posXdeCarta + (anchoCarta / 2) };
 			int[] yPuntos = new int[] { posYdeCarta + (alturaCarta / 2), posYdeCarta + (alturaCarta / 2) - 18,
@@ -672,6 +695,7 @@ public class BlackjackDemo {
 
 		} else if (paloCarta.equals("Corazones")) {
 			// Dibujar el símbolo de Corazones
+			color = Colores.RED;
 			int[] xPuntos = new int[] { posXdeCarta + (anchoCarta / 2) - 14, posXdeCarta + (anchoCarta / 2),
 					posXdeCarta + (anchoCarta / 2) + 14, posXdeCarta + (anchoCarta / 2) };
 			int[] yPuntos = new int[] { posYdeCarta + (alturaCarta / 2) + 1, posYdeCarta + (alturaCarta / 2) - 2,
@@ -682,64 +706,32 @@ public class BlackjackDemo {
 			mesa.dibujarOvaloLleno(posXdeCarta + (anchoCarta / 2) - 2, posYdeCarta + (alturaCarta / 2) - 12, 17, 17,
 					color);
 		}
-	}
 
-	/**
-	 * Metodo con el que se dibuja la forma rectangular de la carta en un color blanco
-	 * y, a su vez, se dibuja el rectangulo interior que delimita el texto de la carta.
-	 * Tambien con este metodo se escribira el nombre de la carta en las esquinas superior
-	 * izquierda e inferior derecha. El nombre de la carta se refiere al "A", 2, 5, "J", "K", etc
-	 */
-	//@ requires mesa != null;
-	//@ requires 0 < mesa.XMAX < Integer.MAX_VALUE;
-	//@ requires 0 < mesa.YMAX < Integer.MAX_VALUE;
-	//@ requires 0 <= (posXdeCarta + 4) < Integer.MAX_VALUE;
-	//@ requires 0 <= (posYdeCarta + 5) < Integer.MAX_VALUE;
-	//@ requires 30 <= alturaCarta <= 10000;
-	//@ requires 30 <= anchoCarta <= 10000;
-	//@ requires (posXdeCarta + anchoCarta) < Integer.MAX_VALUE;
-	//@ requires (posYdeCarta + alturaCarta) < Integer.MAX_VALUE;
-	//@ requires (posXdeCarta + (anchoCarta / 2) + 25) < Integer.MAX_VALUE;
-	//@ requires (posYdeCarta + (alturaCarta / 2) + 18) < Integer.MAX_VALUE;
-	//@ requires nombreCarta.equals("A") || nombreCarta.equals("J") || nombreCarta.equals("Q") || nombreCarta.equals("K") || nombreCarta.equals("Joker") || nombreCarta.equals("2") || nombreCarta.equals("3") || nombreCarta.equals("4") || nombreCarta.equals("5") || nombreCarta.equals("6") || nombreCarta.equals("7") || nombreCarta.equals("8") || nombreCarta.equals("9") || nombreCarta.equals("10");
-	//@ requires color == Colores.BLACK || color == Colores.RED;
-	public static /*@ pure @*/ void dibujarCartaConNombre(MaquinaDeTrazados mesa, int posXdeCarta, int posYdeCarta,
-			int alturaCarta, int anchoCarta, String nombreCarta, Colores color) {
-		// Dibujar el rectángulo externo e interno de la carta
-		mesa.dibujarRectanguloLleno(posXdeCarta, posYdeCarta, 72, 108, Colores.WHITE);
-		mesa.dibujarRectangulo(posXdeCarta + 4, posYdeCarta + 5, anchoCarta - 10, alturaCarta - 10, Colores.DARK_GRAY);
+		// Escribir el nombre de la carta en las esquinas
+		int posXEsqInfIzquierda = posXdeCarta + anchoCarta - 24;
+		int posYEsqInfIzquierda = posYdeCarta + alturaCarta - 12;
 
-		// Configuramos la fuente apropiadamente
 		mesa.configurarFuente("SansSerif", Font.BOLD, 18);
-
-		// Escribimos el nombre de la carta en la esquina superior izquierda
 		mesa.dibujarString(nombreCarta, posXdeCarta + 11, posYdeCarta + 24, color);
-
-		// Ahora determinamos la posicion adecuada del nombre de la carta
-		// que aparecera en la esquina inferior derecha
-		int posXEsqInfDerecha = posXdeCarta + anchoCarta - 24;
-		int posYEsqInfDerecha = posYdeCarta + alturaCarta - 12;
 		if (nombreCarta.equals("A")) {
-			posXEsqInfDerecha = posXdeCarta + anchoCarta - 26;
-			posYEsqInfDerecha = posYdeCarta + alturaCarta - 10;
+			posXEsqInfIzquierda = posXdeCarta + anchoCarta - 26;
+			posYEsqInfIzquierda = posYdeCarta + alturaCarta - 10;
 		} else if (nombreCarta.equals("10")) {
-			posXEsqInfDerecha = posXdeCarta + anchoCarta - 36;
-			posYEsqInfDerecha = posYdeCarta + alturaCarta - 10;
+			posXEsqInfIzquierda = posXdeCarta + anchoCarta - 36;
+			posYEsqInfIzquierda = posYdeCarta + alturaCarta - 10;
 		} else if (nombreCarta.equals("J")) {
-			posXEsqInfDerecha = posXdeCarta + anchoCarta - 17;
-			posYEsqInfDerecha = posYdeCarta + alturaCarta - 13;
+			posXEsqInfIzquierda = posXdeCarta + anchoCarta - 17;
+			posYEsqInfIzquierda = posYdeCarta + alturaCarta - 13;
 		} else if (nombreCarta.equals("Q"))
-			posXEsqInfDerecha = posXdeCarta + anchoCarta - 26;
+			posXEsqInfIzquierda = posXdeCarta + anchoCarta - 26;
 		else if (nombreCarta.equals("K")) {
-			posXEsqInfDerecha = posXdeCarta + anchoCarta - 26;
-			posYEsqInfDerecha = posYdeCarta + alturaCarta - 10;
+			posXEsqInfIzquierda = posXdeCarta + anchoCarta - 26;
+			posYEsqInfIzquierda = posYdeCarta + alturaCarta - 10;
 		} else if (nombreCarta.equals("Joker"))
-			posXEsqInfDerecha = posXdeCarta + 11;
+			posXEsqInfIzquierda = posXdeCarta + 11;
 		else
-			posYEsqInfDerecha = posYdeCarta + alturaCarta - 10;
-
-		// Finalmente escribimos el nombre de la carta en la esquina superior derecha
-		mesa.dibujarString(nombreCarta, posXEsqInfDerecha, posYEsqInfDerecha, color);
+			posYEsqInfIzquierda = posYdeCarta + alturaCarta - 10;
+		mesa.dibujarString(nombreCarta, posXEsqInfIzquierda, posYEsqInfIzquierda, color);
 	}
 
 	/**
@@ -831,65 +823,9 @@ public class BlackjackDemo {
 		mesa.dibujarPoligonoLleno(xPuntos7, yPuntos7, 4, color);
 	}
 
-	/**
-	 * Metodo con el que se mostrara graficamente los valores de las manos
-	 * del crupier y del jugador en la esquina inferior izquierda
-	 */
-	//@ requires mesa != null;
-	//@ requires 0 < mesa.XMAX < Integer.MAX_VALUE;
-	//@ requires 0 < mesa.YMAX < Integer.MAX_VALUE;
-	//@ requires 0 <= alturaMesa < Integer.MAX_VALUE;
-	//@ requires 1 <= valorManoCrupier <= 31;
-	//@ requires 2 <= valorManoJugador <= 31;
-	public static /*@ pure @*/ void mostrarPuntuaciones(MaquinaDeTrazados mesa, int alturaMesa, boolean mostrarCartasCrupier,
-            int valorManoCrupier, int valorManoJugador, Carta[] crupier) {
-		// Convertimos las puntuaciones de int a String para poder escribirlas
-		// en la interfaz grafica
-        if (mostrarCartasCrupier == false)
-			valorManoCrupier = obtenerValorCarta(crupier[0]);
-		String puntuacionCrupier = Integer.toString(valorManoCrupier);
-		String puntuacionJugador = Integer.toString(valorManoJugador);
-
-		// Configuramos la fuente apropiadamente
-		mesa.configurarFuente("SansSerif", Font.BOLD, 22);
-
-		// En el caso que la mano aun no haya terminado y no se hayan revelado
-		// todas las cartas del crupier solamente podemos mostrar el valor
-		// de la primera carta del crupier
-		if (mostrarCartasCrupier == false)
-			mesa.dibujarString("Valor carta visible del crupier: " + puntuacionCrupier, 30, alturaMesa - 105, Colores.BLACK);
-		else
-			mesa.dibujarString("Valor mano del crupier: " + puntuacionCrupier, 30, alturaMesa - 105, Colores.BLACK);
-
-		// Finalmente mostramos el valor de la mano del jugador
-		mesa.dibujarString("Valor mano del jugador: " + puntuacionJugador, 30, alturaMesa - 55, Colores.RED);
-	}
-
-	/**
-	 * Metodo con el que se mostrara graficamente la cantidad de creditos
-	 * que posee el jugador y la cantidad de creditos que se aposto en la mano.
-	 * Estas se muestran en la esquina inferior derecha
-	 */
-	//@ requires mesa != null;
-	//@ requires 0 < mesa.XMAX < Integer.MAX_VALUE;
-	//@ requires 0 < mesa.YMAX < Integer.MAX_VALUE;
-	//@ requires 0 <= alturaMesa < Integer.MAX_VALUE;
-	//@ requires 0 <= anchoMesa < Integer.MAX_VALUE;
-	//@ requires 0 <= credito <= 1000000;
-	//@ requires 10 <= apuesta <= 10000;
-	public static /*@ pure @*/ void mostrarCreditos(MaquinaDeTrazados mesa, int anchoMesa, int alturaMesa,
-			int credito, int apuesta) {
-		String creditosTotales = Integer.toString(credito);
-		String creditosApostados = Integer.toString(apuesta);
-		mesa.configurarFuente("SansSerif", Font.BOLD, 22);
-		mesa.dibujarString("Total de creditos restantes: " + creditosTotales, anchoMesa - 425, alturaMesa - 105, Colores.BLACK);
-		mesa.dibujarString("Creditos apostados: " + creditosApostados, anchoMesa - 325, alturaMesa - 55, Colores.RED);
-	}
-
-	//@ ensures \result != null;
-	//@ ensures \result.length == 56;
-	public static /*@ pure @*/ Carta[] repartirCartasEnElMazo() {
-		Carta[] mazo = new Carta[] {
+    public static void main(String[] args) {
+        // Se crea un mazo con todas las 56 posibles cartas
+        Carta mazo[] = new Carta[] {
             Carta.AS_PICAS, Carta.DOS_PICAS, Carta.TRES_PICAS, Carta.CUATRO_PICAS,
             Carta.CINCO_PICAS, Carta.SEIS_PICAS, Carta.SIETE_PICAS, Carta.OCHO_PICAS,
             Carta.NUEVE_PICAS, Carta.DIEZ_PICAS, Carta.J_PICAS, Carta.Q_PICAS,
@@ -905,99 +841,22 @@ public class BlackjackDemo {
             Carta.SIETE_CORAZON, Carta.OCHO_CORAZON, Carta.NUEVE_CORAZON, Carta.DIEZ_CORAZON,
             Carta.J_CORAZON, Carta.Q_CORAZON, Carta.K_CORAZON, Carta.COMODÍN_CORAZON
         };
-		return mazo;
-	}
-
-	public static void dibujarCartasCrupier(MaquinaDeTrazados mesa, int anchoMesa, int anchoCarta, int alturaCarta, int[] sizes,
-			int numCartasCrupier, Carta[] crupier, boolean mostrarCartasCrupier) {
-		int i = 0;
-		int j = 0;
-		while (i < numCartasCrupier) {
-			int[] posicionesXdeCartas = new int[sizes[0]];
-			int[] posicionesXdeCartas2 = new int[sizes[1]];
-			posicionesXdeCartas = posicionesManoCarta(sizes[0], anchoCarta, anchoMesa);
-			posicionesXdeCartas2 = posicionesManoCarta(sizes[1], anchoCarta, anchoMesa);
-			String paloCarta = obtenerPaloCarta(crupier[i]);
-			Colores color = determinarColorCarta(paloCarta);
-			String nombreCarta = obtenerNombreCarta(crupier[i]);
-			if (mostrarCartasCrupier == false) {
-				if (i == 0) {
-					dibujarCartaConNombre(mesa, posicionesXdeCartas[i], 10, alturaCarta, anchoCarta, nombreCarta, color);
-					dibujarPaloCarta(mesa, posicionesXdeCartas[i], 10, alturaCarta, anchoCarta, paloCarta, color);
-				} else if (i < 12)
-					dibujarCartaVolteada(mesa, posicionesXdeCartas[i], 10, alturaCarta, anchoCarta);
-				else {
-					dibujarCartaVolteada(mesa, posicionesXdeCartas2[j], 128, alturaCarta, anchoCarta);
-					j = j + 1;
-				}
-			} else {
-				if (i < 12) {
-					dibujarCartaConNombre(mesa, posicionesXdeCartas[i], 10, alturaCarta, anchoCarta, nombreCarta, color);
-					dibujarPaloCarta(mesa, posicionesXdeCartas[i], 10, alturaCarta, anchoCarta, paloCarta, color);
-				} else {
-					dibujarCartaConNombre(mesa, posicionesXdeCartas2[j], 128, alturaCarta, anchoCarta, nombreCarta, color);
-					dibujarPaloCarta(mesa, posicionesXdeCartas2[j], 128, alturaCarta, anchoCarta, paloCarta, color);
-					j = j + 1;
-				}
-			}
-			i = i + 1;
-		}
-
-	}
-
-	public static void dibujarCartasJugador(MaquinaDeTrazados mesa, int anchoMesa, int anchoCarta, int alturaCarta, int[] sizes,
-			int numCartasJugador, Carta[] jugador) {
-		int i, j, k;
-		i = j = k = 0;
-		while (i < numCartasJugador) {
-			int[] posicionesXdeCartas = new int[sizes[0]];
-			int[] posicionesXdeCartas2 = new int[sizes[1]];
-			int[] posicionesXdeCartas3 = new int[sizes[2]];
-			posicionesXdeCartas = posicionesManoCarta(sizes[0], anchoCarta, anchoMesa);
-			posicionesXdeCartas2 = posicionesManoCarta(sizes[1], anchoCarta, anchoMesa);
-			posicionesXdeCartas3 = posicionesManoCarta(sizes[2], anchoCarta, anchoMesa);
-			String paloCarta = obtenerPaloCarta(jugador[i]);
-			Colores color = determinarColorCarta(paloCarta);
-			String nombreCarta = obtenerNombreCarta(jugador[i]);
-			if (i < 10) {
-				dibujarCartaConNombre(mesa, posicionesXdeCartas[i], 256, alturaCarta, anchoCarta, nombreCarta, color);
-				dibujarPaloCarta(mesa, posicionesXdeCartas[i], 256, alturaCarta, anchoCarta, paloCarta, color);
-			} else if (i < 18) {
-				dibujarCartaConNombre(mesa, posicionesXdeCartas2[j], 374, alturaCarta, anchoCarta, nombreCarta, color);
-				dibujarPaloCarta(mesa, posicionesXdeCartas2[j], 374, alturaCarta, anchoCarta, paloCarta, color);
-				j = j + 1;
-			} else {
-				dibujarCartaConNombre(mesa, posicionesXdeCartas3[k], 492, alturaCarta, anchoCarta, nombreCarta, color);
-				dibujarPaloCarta(mesa, posicionesXdeCartas3[k], 492, alturaCarta, anchoCarta, paloCarta, color);
-				k = k + 1;
-			}
-			i = i + 1;
-		}
-	}
-
-    public static /*@ pure @*/ void mensajeDespedida(String nombre) {
-        System.out.println("");
-        System.out.println("Termino el juego. Ha sido un placer tenerte aqui, " + nombre + ".");
-        System.out.println("Gracias por jugar y regresa pronto.");
-    }
-
-    public static void main(String[] args) {
-        // Se crea un mazo con todas las 56 posibles cartas
-        Carta mazo[] = new Carta[56];
-		mazo = repartirCartasEnElMazo();
 
         // Se declaran e inicializan las variables basicas que se usaran en el programa
         Console con = System.console();
         Carta jugador[] = new Carta[21];
         Carta crupier[] = new Carta[17];
         int credito = 100;
-		int apuesta, puntosJugador, puntosCrupier, numManosJugadas;
-        apuesta = puntosJugador = puntosCrupier = numManosJugadas = 0;
-		int numCartasCrupier = 2;
-		int numCartasJugador = 2;
-        boolean seJugaraOtraMano, mostrarCartasCrupier, finalizoLaMano, mostrarResultado, hayBlackjack, sePasaronDeVeintiuno, resultadoMostrado, calcularResultadoMano, terminoElTurnoDelJugador;
-        seJugaraOtraMano = mostrarCartasCrupier = finalizoLaMano = mostrarResultado = resultadoMostrado = hayBlackjack = sePasaronDeVeintiuno = calcularResultadoMano = terminoElTurnoDelJugador = false;
-        boolean acabaDeEmpezarLaMano = true;
+		int apuesta = 0;
+        int puntosJugador = 0;
+        int puntosCrupier = 0;
+		int opcion = 0;
+		int mensaje = 0;
+		int comprobante = 0;
+		int cota_FM = 0;
+		int maxJugadas = 0;
+		boolean salir_juego = false;
+		boolean fin_mano = false;
 
         // Se especifican las dimensiones de las cartas
 		int alturaCarta = 108;
@@ -1006,152 +865,86 @@ public class BlackjackDemo {
 		int alturaMesa = 774;
 		int anchoMesa = 1278;
 
-        // Mensaje de bienvenida donde se le pregunta el nombre al jugador
-		String nombre = mensajeBienvenida();
-
-		// Se crea el panel de la Máquina de Trazados. El color gris crea un fondo agradable
+		// Se crea el panel de la Maquina de Trazados. El color gris crea un fondo agradable
 		MaquinaDeTrazados mesa = new MaquinaDeTrazados(anchoMesa, alturaMesa, "BlackJack", Colores.GRAY);
 
-		while (credito >= 0 && numManosJugadas < 5) {
-            // Empieza una mano
-            if (acabaDeEmpezarLaMano == true) {
-                // Se le pide al jugador que ingrese el monto de la apuesta usando su nombre
-                // Tambien se le explica al jugador el sistema de creditos y apuestas
-                apuesta = obtenerApuesta(credito, nombre);
-                credito = calcularCreditoResultante(apuesta, credito, 0);
+        // Semicirculo verde que imita una mesa de BlackJack real con un borde oscuro
+        mesa.dibujarOvaloLleno(1, -(anchoMesa / 2), anchoMesa, anchoMesa, Colores.DARK_GRAY);
+        mesa.dibujarOvaloLleno(11, -(anchoMesa / 2), anchoMesa - 20, anchoMesa - 20, Colores.GREEN);
 
-                // Una vez se tiene el monto que aposto el jugador, se reparten las cartas al azar
-                // Pero solo la primera vez que se itera en la misma mano, sino se cambiarian
-                // las cartas cuando se pida una nueva
-                jugador = repartirCartasIniciales(jugador, mazo);
-                crupier = repartirCartasIniciales(crupier, mazo);
-            }
+        // Mensaje de bienvenida donde se le pregunta el nombre al jugador
+		
+		String nombre = mensajeEntrada();
 
-			// Dibujar la mesa de blackjack
-			dibujarMesaBlackjack(mesa, anchoMesa, alturaMesa, alturaCarta, anchoCarta);
+        // Se le explica al jugador el sistema de creditos y apuestas
+        // Luego se le pide al jugador que ingrese el monto de la apuesta usando su nombre
+        System.out.println("Posees 100 creditos por empezar a jugar. La apuesta minima por mano son 10 creditos.");
+        
+		while (!salir_juego && credito >= 10 && maxJugadas != 5) {
+			apuesta = obtenerApuesta(credito, nombre);
+			credito = operacionCredito(apuesta, credito, 1);
 
-			// Dibujar las cartas del crupier. Mientras el jugador no se haya
-            // platado, se dibuja la primera carta visible y el resto volteadas
-            // Luego, cuando el jugador se planta, se dibujan todas las cartas
-            // del crupier de forma visible
-			int[] sizes = new int[3];
-			sizes = getSizes(numCartasCrupier, "crupier");
-			dibujarCartasCrupier(mesa, anchoMesa, anchoCarta, alturaCarta, sizes, numCartasCrupier, crupier, mostrarCartasCrupier);
+			// Una vez se tiene el monto que aposto el jugador, se reparten las cartas al azar
+			jugador = primeras2Cartas(jugador, mazo);
+			crupier = primeras2Cartas(crupier, mazo);
 
-			// Inmediatamente, se dibujan las cartas del jugador, todas visibles
-			sizes = getSizes(numCartasJugador, "jugador");
-			dibujarCartasJugador(mesa, anchoMesa, anchoCarta, alturaCarta, sizes, numCartasJugador, jugador);
+			// Luego se dibujan las cartas del crupier
+			dibujarCartaVisible(mesa, (anchoMesa / 2) - 82, 50, alturaCarta, anchoCarta, obtenerPaloCarta(crupier[0]), obtenerNombreCarta(crupier[0]));
+			dibujarCartaVolteada(mesa, (anchoMesa / 2) + 10, 50, alturaCarta, anchoCarta);
 
-			// Calculamos el valor de las manos del crupier y del jugador
-			puntosCrupier = valorMano(crupier, numCartasCrupier);
-			puntosJugador = valorMano(jugador, numCartasJugador);
-
-			// Mostramos las puntuaciones y los creditos en la interfaz
-			// grafica para informar al jugador
-			mostrarPuntuaciones(mesa, alturaMesa, mostrarCartasCrupier, puntosCrupier, puntosJugador, crupier);
-			mostrarCreditos(mesa, anchoMesa, alturaMesa, credito, apuesta);
-
+			// E inmediatamente se dibujan las cartas del jugador
+			dibujarCartaVisible(mesa, (anchoMesa / 2) - 82, 300, alturaCarta, anchoCarta, obtenerPaloCarta(jugador[0]),  obtenerNombreCarta(jugador[0]));
+			dibujarCartaVisible(mesa, (anchoMesa / 2) + 10, 300, alturaCarta, anchoCarta, obtenerPaloCarta(jugador[1]),  obtenerNombreCarta(jugador[1]));
 			mesa.mostrar();
 
-			// Se le muestra al jugador la informacion de la mano por la
-            // interfaz de texto tambien
-			mostrarPuntuacionesPorTexto(puntosCrupier, puntosJugador, mostrarCartasCrupier, crupier);
-            mostrarCreditosPorTexto(credito, apuesta);
+			// Calculamos el valor de las manos en la mesa
+			puntosJugador = valorMano(jugador, 2, puntosJugador);
+			puntosCrupier = valorMano(crupier, 2, puntosCrupier);
 
-            // Si termina la mano se le muestra el resultado al jugador hasta
-            // que el jugador introduzca un caracter
-            if (mostrarResultado == true) {
-                con.readLine("Ingrese cualquier caracter para aceptar el resultado de la mano: ");
-                // Variable necesaria para mostrar las cartas del crupier
-                // justo cuando termine la mano
-                mostrarResultado = false;
-                resultadoMostrado = true;
-            }
+			// Se le muestra al jugador el valor de las cartas en la mesa en la interfaz grafica
+			mesa.configurarFuente("SansSerif", Font.BOLD, 18);
+			mesa.dibujarString("Carta descubierta del crupier: " + Integer.toString(obtenerValorCarta(crupier[0])), (anchoMesa / 2) - 160, 30, Colores.DARK_GRAY);
+			mesa.dibujarString("Mano actual: " + Integer.toString(puntosJugador), (anchoMesa / 2) - 82, 280, Colores.DARK_GRAY);
 
-            // Se revisa si el jugador o el crupier hicieron blackjack
-            // 0 si el jugador o el crupier se pasaron de 21
-            hayBlackjack = revisarSiHayBlackjack(puntosCrupier, puntosJugador);
-            sePasaronDeVeintiuno = revisarSiSePasaronDeVeintiuno(puntosCrupier, puntosJugador);
-            if ((hayBlackjack == true || sePasaronDeVeintiuno == true) && resultadoMostrado == false) {
-                credito = determinarResultadoMano(apuesta, credito, puntosCrupier, puntosJugador, nombre);
-                finalizoLaMano = mostrarCartasCrupier = mostrarResultado = true;
-            }
+			// Se le muestra al jugador en la interfaz grafica el numero de creditos totales
+			// que posee y tambien cuantos se apostaron en la mano actual
+			mesa.dibujarString("Creditos actuales: " + Integer.toString(credito), (anchoMesa / 2) - 410, 210, Colores.BLUE);
+			mesa.dibujarString("Creditos apostados: " + Integer.toString(apuesta), (anchoMesa / 2) + 190, 210, Colores.RED);
 
-            // Si no hubo blackjack ni se pasaron de 21, se le da a elegir
-            // al jugador que hacer a continuacion
-            if (sePasaronDeVeintiuno == false && hayBlackjack == false && resultadoMostrado == false) {
-                // A continuacion se le solicita al jugador decidir si se planta,
-                // si pide una carta adicional, si dobla su apuesta (cuando las
-                // condiciones lo permiten) o si desea salir del juego
-                int decisionJugador = consultarDecisionJugador(nombre, puntosJugador, acabaDeEmpezarLaMano, credito, apuesta);
-                // Con el valor false de esta variable se impide que se vuelvan a repartir
-                // cartas al jugador o al crupier. De esa forma es seguro pedir una nueva
-                // carta, sabiendo que las anteriores seguiran siendo las mismas
-                acabaDeEmpezarLaMano = false;
-                if (decisionJugador == 1) {
-                    jugador[numCartasJugador] = darCarta(mazo);
-                    numCartasJugador = numCartasJugador + 1;
-                } else if (decisionJugador == 2) {
-                    mostrarCartasCrupier = finalizoLaMano = mostrarResultado = calcularResultadoMano = terminoElTurnoDelJugador = true;
-                } else if (decisionJugador == 3) {
-                    credito = calcularCreditoResultante(apuesta, credito, 0);
-                    apuesta = doblarApuesta(apuesta, credito);
-                    jugador[numCartasJugador] = darCarta(mazo);
-                    numCartasJugador = numCartasJugador + 1;
-                    mostrarCartasCrupier = finalizoLaMano = mostrarResultado = calcularResultadoMano = terminoElTurnoDelJugador = true;
-                }
-                // El jugador decidio salirse del juego
-                else
-                    break;
-
-                // El crupier pedira una carta siempre de el valor de su mano
-                // sea menor igual a 16
-                while (terminoElTurnoDelJugador == true && puntosCrupier <= 16) {
-                    crupier[numCartasCrupier] = darCarta(mazo);
-                    numCartasCrupier = numCartasCrupier + 1;
-                    puntosCrupier = valorMano(crupier, numCartasCrupier);
-                }
-            }
-
-            // Se determina el resultado de la mano luego de plantarse o doblar la apuesta
-            if (calcularResultadoMano == true) {
-                credito = determinarResultadoMano(apuesta, credito, puntosCrupier, puntosJugador, nombre);
-                calcularResultadoMano = false;
-            }
-
-            // Se cierra y se borra el panel de la maquina de trazados
-            // Esto es para poder redibujar el panel en el caso de que el jugador
-            // o el crupier hayan decidido agarrar mas cartas
-			mesa.terminar();
-			mesa.limpiar();
-
-            // Cuando termina la mano se le pregunta al jugador si quiere volver
-            // a jugar otra mano o si, por otro lado, prefiere salirse del juego
-			if (finalizoLaMano == true && resultadoMostrado == true) {
-                // El jugador alcanzo el numero maximo de manos que puede jugar en una
-                // misma sesion, por ello se termina el juego
-                if (credito < 10) {
-                    System.out.println("");
-                    System.out.println("Tienes menos creditos que la apuesta minina para una mano. Termina el juego.");
-                    break;
-                } else if (numManosJugadas == 4) {
-                    System.out.println("");
-                    System.out.println("Has alcanzado la maxima cantidad de manos jugadas por sesion.");
-                    break;
-                } else
-                    seJugaraOtraMano = determinarSiSeJugaraOtraMano(credito);
-                if (seJugaraOtraMano == true) {
-                    numManosJugadas = numManosJugadas + 1;
-                    numCartasCrupier = numCartasJugador = 2;
-                    acabaDeEmpezarLaMano = true;
-                    finalizoLaMano = mostrarCartasCrupier = mostrarResultado = resultadoMostrado = false;
-                } else
-                    break;
-            }
+			// Se le muestra al jugador el valor de las cartas en la mesa en la interfaz de texto
+			primeraMuestra(puntosJugador, crupier);
+			int i = 2;
+			int j = 2;
+			while (!fin_mano) {
+				int comparar = acciones(puntosJugador, puntosCrupier, mensaje, mensaje, apuesta, credito, opcion, comprobante, mazo, jugador, crupier, nombre);
+				if (comparar == 0){
+					salir_juego = true;
+					break;
+				} else {
+					credito = comparar;
+					break;
+				}
+			}
+			if (salir_juego == true) {
+				break;
+			}
+			cota_FM = end_mano(cota_FM, credito);
+			if (cota_FM == 1){
+				maxJugadas++;
+				puntosJugador = 0;
+        		puntosCrupier = 0;
+				opcion = 0;
+				mensaje = 0;
+				comprobante = 0;
+				cota_FM = 0;
+				fin_mano = false;
+			} else 
+				break; 
+				
+			if (maxJugadas == 5)
+				System.out.println("Haz alcanzado la maxima cantidad de jugadas disponibles");
 		}
-
-		// Se cierra el programa y se despide al jugador
-		mensajeDespedida(nombre);
+		System.out.println("El juego se ha terminado");
 		mesa.terminar();
-	}
+    }
 }
